fix: moved sectiontabitem out, fixed test
