+++ src/components/layout/header/SectionTab.tsx
@@ -1,26 +1,5 @@
-import { cn } from 'Utils/cn';
 import { useTranslation } from 'react-i18next';
-
-type TSectionTabItemProps = {
-    alt: string;
-    className?: string;
-    imageUrl: string;
-    label: string;
-};
-
-const SectionTabItem = ({ alt, label, imageUrl, className }: TSectionTabItemProps) => {
-    return (
-        <a
-            className={cn(
-                'flex h-full cursor-pointer flex-row items-center gap-1.5 px-[0.5rem] hover:bg-disabled-100',
-                className
-            )}
-        >
-            <img src={imageUrl} alt={alt} />
-            {label}
-        </a>
-    );
-};
+import SectionTabItem from './SectionTabItem';
 
 const SectionTab = () => {
     const { t } = useTranslation();
