+++ src/components/layout/header/__test__/LoggedOutState.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import LoggedInState from '../LoggedInState';
 
 describe('LoggedInState', () => {
-    it('should render the logged out state', () => {
+    it('should render the logged in state', () => {
         render(<LoggedInState />);
         expect(screen.getByText('Deposit')).toBeInTheDocument();
     });
